Reworked whole search engine
Suddenly search engine wasnt working. I wasnt able to find the mistake so i reworked the whole thing to perfection. When user searches for an example first name and there is only one registered account with that username, the engine automaticly opens the profile, if there is more, list of users with the name pops up.